requirements.txt
langchain==0.1.20
langchain-community==0.0.38
streamlit==1.32.0
ollama==0.1.7
textblob==0.17.1
pandas==2.2.0
plotly==5.19.0
python-dotenv==1.0.1
bcrypt==4.1.2
flask==3.0.0
flask-cors==4.0.0
supabase==2.0.0
